MINOR: Move inner test classes out of CoordinatorRuntimeTest
Some of these classes are generally useful for testing.
MockCoordinatorShard is already shared by SnapshottableCoordinatorTest.
