+++ coordinator-common/src/test/java/org/apache/kafka/coordinator/common/runtime/ThrowingSerializer.java
@@ -1,0 +1,50 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.kafka.coordinator.common.runtime;
+
+import java.nio.BufferOverflowException;
+
+/**
+ * A Serializer that throws exceptions on demand.
+ */
+public class ThrowingSerializer<T> implements Serializer<T> {
+    private final Serializer<T> serializer;
+    private boolean throwOnNextOperation;
+
+    public ThrowingSerializer(Serializer<T> serializer) {
+        this.serializer = serializer;
+        this.throwOnNextOperation = false;
+    }
+
+    public void throwOnNextOperation() {
+        throwOnNextOperation = true;
+    }
+
+    @Override
+    public byte[] serializeKey(T record) {
+        return serializer.serializeKey(record);
+    }
+
+    @Override
+    public byte[] serializeValue(T record) {
+        if (throwOnNextOperation) {
+            throwOnNextOperation = false;
+            throw new BufferOverflowException();
+        }
+        return serializer.serializeValue(record);
+    }
+}
